Put primary type at first for node property type

+++ gdcdatamodel/models/nodes/archive.py
@@ -9,6 +9,6 @@ class Archive(Node):
     def submitter_id(self, value):
         self._set_property('submitter_id', value)
 
-    @pg_property(int, long)
+    @pg_property(long, int)
     def revision(self, value):
         self._set_property('revision', value)
